Move SierraRecord into the sierra_adapter common lib
This model is internal to the Sierra adapter pipeline -- by the time it
reaches the VHS, it's been coerced into a BibRecord or an ItemRecord.
It doesn't need to be available to the transformer or other applications.
