+++ sierra_adapter/common/src/main/scala/uk/ac/wellcome/sierra_adapter/models/SierraRecord.scala
@@ -1,10 +1,11 @@
-package uk.ac.wellcome.models.transformable.sierra
+package uk.ac.wellcome.sierra_adapter.models
 
 import java.time.Instant
 
 import io.circe.optics.JsonPath.root
 import io.circe.parser._
 import cats.syntax.either._
+import uk.ac.wellcome.models.transformable.sierra.{SierraBibRecord, SierraItemRecord}
 
 import scala.util.Try
