+++ sierra_adapter/sierra_bib_merger/src/main/scala/uk/ac/wellcome/platform/sierra_bib_merger/services/SierraBibMergerWorkerService.scala
@@ -4,11 +4,11 @@ import akka.actor.ActorSystem
 import com.google.inject.Inject
 import io.circe.Decoder
 import uk.ac.wellcome.messaging.sqs.{SQSReader, SQSWorkerToDynamo}
-import uk.ac.wellcome.models.transformable.sierra.SierraRecord
 import uk.ac.wellcome.utils.JsonUtil._
 import uk.ac.wellcome.utils.GlobalExecutionContext.context
 import io.circe.generic.extras.semiauto._
 import uk.ac.wellcome.monitoring.MetricsSender
+import uk.ac.wellcome.sierra_adapter.models.SierraRecord
 
 import scala.concurrent.Future
